Move logic from SignednessVisitor to SignednessAnnotatedTypeFactory
