+++ checker/src/main/java/org/checkerframework/checker/signedness/SignednessAnnotatedTypeFactory.java
@@ -1,14 +1,22 @@
 package org.checkerframework.checker.signedness;
 
+import com.sun.source.tree.AnnotatedTypeTree;
 import com.sun.source.tree.BinaryTree;
 import com.sun.source.tree.CompoundAssignmentTree;
+import com.sun.source.tree.ExpressionTree;
+import com.sun.source.tree.LiteralTree;
+import com.sun.source.tree.PrimitiveTypeTree;
 import com.sun.source.tree.Tree;
+import com.sun.source.tree.Tree.Kind;
+import com.sun.source.tree.TypeCastTree;
+import com.sun.source.util.TreePath;
 import java.lang.annotation.Annotation;
 import java.util.Set;
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.Element;
 import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
+import org.checkerframework.checker.interning.qual.InternedDistinct;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.checker.signedness.qual.Signed;
 import org.checkerframework.checker.signedness.qual.SignedPositive;
@@ -31,6 +39,9 @@ import org.checkerframework.framework.type.treeannotator.PropagationTreeAnnotato
 import org.checkerframework.framework.type.treeannotator.TreeAnnotator;
 import org.checkerframework.javacutil.AnnotationBuilder;
 import org.checkerframework.javacutil.AnnotationUtils;
+import org.checkerframework.javacutil.BugInCF;
+import org.checkerframework.javacutil.Pair;
+import org.checkerframework.javacutil.TreeUtils;
 import org.checkerframework.javacutil.TypeKindUtils;
 import org.checkerframework.javacutil.TypesUtils;
 
@@ -241,8 +252,15 @@ public class SignednessAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
                 case LEFT_SHIFT:
                 case RIGHT_SHIFT:
                 case UNSIGNED_RIGHT_SHIFT:
-                    AnnotatedTypeMirror lht = getAnnotatedType(tree.getLeftOperand());
-                    type.replaceAnnotations(lht.getAnnotations());
+                    TreePath path = getPath(tree);
+                    if (path != null
+                            && (isMaskedShiftEitherSignedness(tree, path)
+                                    || isCastedShiftEitherSignedness(tree, path))) {
+                        type.replaceAnnotation(SIGNEDNESS_GLB);
+                    } else {
+                        AnnotatedTypeMirror lht = getAnnotatedType(tree.getLeftOperand());
+                        type.replaceAnnotations(lht.getAnnotations());
+                    }
                     break;
                 default:
                     // Do nothing
@@ -270,5 +288,262 @@ public class SignednessAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
         } else {
             super.addAnnotationsFromDefaultForType(element, type);
         }
+    }
+
+    // Special case for shifts whose relust does not depend on the MSB of the first argument.
+
+    /** @return true iff node is a mask operation (&amp; or |) */
+    private boolean isMask(Tree node) {
+        Kind kind = node.getKind();
+
+        return kind == Kind.AND || kind == Kind.OR;
+    }
+
+    // TODO: Return a TypeKind rather than a PrimitiveTypeTree?
+    /**
+     * Returns the type of a primitive cast, or null the argument is not a cast to a primitive.
+     *
+     * @param node a tree that might be a cast to a primitive
+     * @return type of a primitive cast, or null if not a cast to a primitive.
+     */
+    private PrimitiveTypeTree primitiveTypeCast(Tree node) {
+        if (node.getKind() != Kind.TYPE_CAST) {
+            return null;
+        }
+
+        TypeCastTree cast = (TypeCastTree) node;
+        Tree castType = cast.getType();
+
+        Tree underlyingType;
+        if (castType.getKind() == Kind.ANNOTATED_TYPE) {
+            underlyingType = ((AnnotatedTypeTree) castType).getUnderlyingType();
+        } else {
+            underlyingType = castType;
+        }
+
+        if (underlyingType.getKind() != Kind.PRIMITIVE_TYPE) {
+            return null;
+        }
+
+        return (PrimitiveTypeTree) underlyingType;
+    }
+
+    /** @return true iff expr is a literal */
+    private boolean isLiteral(ExpressionTree expr) {
+        return expr instanceof LiteralTree;
+    }
+
+    /**
+     * @param obj either an Integer or a Long
+     * @return the long value of obj
+     */
+    private long getLong(Object obj) {
+        return ((Number) obj).longValue();
+    }
+
+    /**
+     * Given a masking operation of the form {@code expr & maskLit} or {@code expr | maskLit},
+     * return true iff the masking operation results in the same output regardless of the value of
+     * the shiftAmount most significant bits of expr. This is if the shiftAmount most significant
+     * bits of mask are 0 for AND, and 1 for OR. For example, assuming that shiftAmount is 4, the
+     * following is true about AND and OR masks:
+     *
+     * <p>{@code expr & 0x0[anything] == 0x0[something] ;}
+     *
+     * <p>{@code expr | 0xF[anything] == 0xF[something] ;}
+     *
+     * @param maskKind the kind of mask (AND or OR)
+     * @param shiftAmountLit the LiteralTree whose value is shiftAmount
+     * @param maskLit the LiteralTree whose value is mask
+     * @param shiftedTypeKind the type of shift operation; int or long
+     * @return true iff the shiftAmount most significant bits of mask are 0 for AND, and 1 for OR
+     */
+    private boolean maskIgnoresMSB(
+            Kind maskKind,
+            LiteralTree shiftAmountLit,
+            LiteralTree maskLit,
+            TypeKind shiftedTypeKind) {
+        long shiftAmount = getLong(shiftAmountLit.getValue());
+
+        // Shift of zero is a nop
+        if (shiftAmount == 0) {
+            return true;
+        }
+
+        long mask = getLong(maskLit.getValue());
+        // Shift the shiftAmount most significant bits to become the shiftAmount least significant
+        // bits, zeroing out the rest.
+        if (shiftedTypeKind == TypeKind.INT) {
+            mask <<= 32;
+        }
+        mask >>>= (64 - shiftAmount);
+
+        if (maskKind == Kind.AND) {
+            // Check that the shiftAmount most significant bits of the mask were 0.
+            return mask == 0;
+        } else if (maskKind == Kind.OR) {
+            // Check that the shiftAmount most significant bits of the mask were 1.
+            return mask == (1 << shiftAmount) - 1;
+        } else {
+            throw new BugInCF("Invalid Masking Operation");
+        }
+    }
+
+    /**
+     * Given a casted right shift of the form {@code (type) (baseExpr >> shiftAmount)} or {@code
+     * (type) (baseExpr >>> shiftAmount)}, return true iff the expression's value is the same
+     * regardless of the type of right shift (signed or unsigned). This is true if the cast ignores
+     * the shiftAmount most significant bits of the shift result -- that is, if the cast ignores all
+     * the new bits that the right shift introduced on the left.
+     *
+     * <p>For example, the function returns true for
+     *
+     * <pre>{@code (short) (myInt >> 16)}</pre>
+     *
+     * and for
+     *
+     * <pre>{@code (short) (myInt >>> 16)}</pre>
+     *
+     * because these two expressions are guaranteed to have the same result.
+     *
+     * @param shiftTypeKind the kind of the type of the shift literal (BYTE, CHAR, SHORT, INT, or
+     *     LONG)
+     * @param castTypeKind the kind of the cast target type (BYTE, CHAR, SHORT, INT, or LONG)
+     * @param shiftAmountLit the LiteralTree whose value is shiftAmount
+     * @return true iff introduced bits are discarded
+     */
+    private boolean castIgnoresMSB(
+            TypeKind shiftTypeKind, TypeKind castTypeKind, LiteralTree shiftAmountLit) {
+
+        // Determine number of bits in the shift type, note shifts upcast to int.
+        // Also determine the shift amount as it is dependent on the shift type.
+        long shiftBits;
+        long shiftAmount;
+        switch (shiftTypeKind) {
+            case INT:
+                shiftBits = 32;
+                // When the LHS of the shift is an int, the 5 lower order bits of the RHS are used.
+                shiftAmount = 0x1F & getLong(shiftAmountLit.getValue());
+                break;
+            case LONG:
+                shiftBits = 64;
+                // When the LHS of the shift is a long, the 6 lower order bits of the RHS are used.
+                shiftAmount = 0x3F & getLong(shiftAmountLit.getValue());
+                break;
+            default:
+                throw new BugInCF("Invalid shift type");
+        }
+
+        // Determine number of bits in the cast type
+        long castBits;
+        switch (castTypeKind) {
+            case BYTE:
+                castBits = 8;
+                break;
+            case CHAR:
+                castBits = 8;
+                break;
+            case SHORT:
+                castBits = 16;
+                break;
+            case INT:
+                castBits = 32;
+                break;
+            case LONG:
+                castBits = 64;
+                break;
+            default:
+                throw new BugInCF("Invalid cast target");
+        }
+
+        long bitsDiscarded = shiftBits - castBits;
+
+        return shiftAmount <= bitsDiscarded || shiftAmount == 0;
+    }
+
+    /**
+     * Determines if a right shift operation, {@code >>} or {@code >>>}, is masked with a masking
+     * operation of the form {@code shiftExpr & maskLit} or {@code shiftExpr | maskLit} such that
+     * the mask renders the shift signedness ({@code >>} vs {@code >>>}) irrelevent by destroying
+     * the bits duplicated into the shift result. For example, the following pairs of right shifts
+     * on {@code byte b} both produce the same results under any input, because of their masks:
+     *
+     * <p>{@code (b >> 4) & 0x0F == (b >>> 4) & 0x0F;}
+     *
+     * <p>{@code (b >> 4) | 0xF0 == (b >>> 4) | 0xF0;}
+     *
+     * @param shiftExpr a right shift expression: {@code expr1 >> expr2} or {@code expr1 >>> expr2}
+     * @return true iff the right shift is masked such that a signed or unsigned right shift has the
+     *     same effect
+     */
+    /*package-private*/ boolean isMaskedShiftEitherSignedness(BinaryTree shiftExpr, TreePath path) {
+        Pair<Tree, Tree> enclosingPair = TreeUtils.enclosingNonParen(path);
+        // enclosing immediately contains shiftExpr or a parenthesized version of shiftExpr
+        Tree enclosing = enclosingPair.first;
+        // enclosingChild is a child of enclosing:  shiftExpr or a parenthesized version of it.
+        @SuppressWarnings("interning:assignment.type.incompatible") // comparing AST nodes
+        @InternedDistinct Tree enclosingChild = enclosingPair.second;
+
+        if (!isMask(enclosing)) {
+            return false;
+        }
+
+        BinaryTree maskExpr = (BinaryTree) enclosing;
+        ExpressionTree shiftAmountExpr = shiftExpr.getRightOperand();
+
+        // Determine which child of maskExpr leads to shiftExpr. The other one is the mask.
+        ExpressionTree mask =
+                maskExpr.getRightOperand() == enclosingChild
+                        ? maskExpr.getLeftOperand()
+                        : maskExpr.getRightOperand();
+
+        // Strip away the parentheses from the mask if any exist
+        mask = TreeUtils.withoutParens(mask);
+
+        if (!isLiteral(shiftAmountExpr) || !isLiteral(mask)) {
+            return false;
+        }
+
+        LiteralTree shiftLit = (LiteralTree) shiftAmountExpr;
+        LiteralTree maskLit = (LiteralTree) mask;
+
+        return maskIgnoresMSB(
+                maskExpr.getKind(), shiftLit, maskLit, TreeUtils.typeOf(shiftExpr).getKind());
+    }
+
+    /**
+     * Determines if a right shift operation, {@code >>} or {@code >>>}, is type casted such that
+     * the cast renders the shift signedness ({@code >>} vs {@code >>>}) irrelevent by discarding
+     * the bits duplicated into the shift result. For example, the following pair of right shifts on
+     * {@code short s} both produce the same results under any input, because of type casting:
+     *
+     * <p>{@code (byte)(s >> 8) == (byte)(b >>> 8);}
+     *
+     * @param shiftExpr a right shift expression: {@code expr1 >> expr2} or {@code expr1 >>> expr2}
+     * @return true iff the right shift is type casted such that a signed or unsigned right shift
+     *     has the same effect
+     */
+    /*package-private*/ boolean isCastedShiftEitherSignedness(BinaryTree shiftExpr, TreePath path) {
+        // enclosing immediately contains shiftExpr or a parenthesized version of shiftExpr
+        Tree enclosing = TreeUtils.enclosingNonParen(path).first;
+
+        PrimitiveTypeTree castPrimitiveType = primitiveTypeCast(enclosing);
+        if (castPrimitiveType == null) {
+            return false;
+        }
+        TypeKind castTypeKind = castPrimitiveType.getPrimitiveTypeKind();
+
+        // Determine the type of the shift result
+        TypeKind shiftTypeKind = TreeUtils.typeOf(shiftExpr).getKind();
+
+        // Determine shift literal
+        ExpressionTree shiftAmountExpr = shiftExpr.getRightOperand();
+        if (!isLiteral(shiftAmountExpr)) {
+            return false;
+        }
+        LiteralTree shiftLit = (LiteralTree) shiftAmountExpr;
+
+        boolean result = castIgnoresMSB(shiftTypeKind, castTypeKind, shiftLit);
+        return result;
     }
 }
