+++ checker/src/main/java/org/checkerframework/checker/signedness/SignednessVisitor.java
@@ -1,25 +1,16 @@
 package org.checkerframework.checker.signedness;
 
-import com.sun.source.tree.AnnotatedTypeTree;
 import com.sun.source.tree.BinaryTree;
 import com.sun.source.tree.CompoundAssignmentTree;
 import com.sun.source.tree.ExpressionTree;
-import com.sun.source.tree.LiteralTree;
-import com.sun.source.tree.PrimitiveTypeTree;
-import com.sun.source.tree.Tree;
 import com.sun.source.tree.Tree.Kind;
-import com.sun.source.tree.TypeCastTree;
-import javax.lang.model.type.TypeKind;
-import org.checkerframework.checker.interning.qual.InternedDistinct;
 import org.checkerframework.checker.signedness.qual.PolySigned;
 import org.checkerframework.checker.signedness.qual.Signed;
 import org.checkerframework.checker.signedness.qual.Unsigned;
 import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.common.basetype.BaseTypeVisitor;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
-import org.checkerframework.javacutil.BugInCF;
 import org.checkerframework.javacutil.Pair;
-import org.checkerframework.javacutil.TreeUtils;
 
 /**
  * The SignednessVisitor enforces the Signedness Checker rules. These rules are described in the
@@ -31,259 +22,6 @@ public class SignednessVisitor extends BaseTypeVisitor<SignednessAnnotatedTypeFa
 
     public SignednessVisitor(BaseTypeChecker checker) {
         super(checker);
-    }
-
-    /** @return true iff node is a mask operation (&amp; or |) */
-    private boolean isMask(Tree node) {
-        Kind kind = node.getKind();
-
-        return kind == Kind.AND || kind == Kind.OR;
-    }
-
-    /**
-     * Returns the type of a primitive cast, or null the argument is not a cast to a primitive.
-     *
-     * @param node a tree that might be a cast to a primitive
-     * @return type of a primitive cast, or null if not a cast to a primitive.
-     */
-    private PrimitiveTypeTree primitiveTypeCast(Tree node) {
-        if (node.getKind() != Kind.TYPE_CAST) {
-            return null;
-        }
-
-        TypeCastTree cast = (TypeCastTree) node;
-        Tree castType = cast.getType();
-
-        Tree underlyingType;
-        if (castType.getKind() == Kind.ANNOTATED_TYPE) {
-            underlyingType = ((AnnotatedTypeTree) castType).getUnderlyingType();
-        } else {
-            underlyingType = castType;
-        }
-
-        if (underlyingType.getKind() != Kind.PRIMITIVE_TYPE) {
-            return null;
-        }
-
-        return (PrimitiveTypeTree) underlyingType;
-    }
-
-    /** @return true iff expr is a literal */
-    private boolean isLiteral(ExpressionTree expr) {
-        return expr instanceof LiteralTree;
-    }
-
-    /**
-     * @param obj either an Integer or a Long
-     * @return the long value of obj
-     */
-    private long getLong(Object obj) {
-        return ((Number) obj).longValue();
-    }
-
-    /**
-     * Given a masking operation of the form {@code expr & maskLit} or {@code expr | maskLit},
-     * return true iff the masking operation results in the same output regardless of the value of
-     * the shiftAmount most significant bits of expr. This is if the shiftAmount most significant
-     * bits of mask are 0 for AND, and 1 for OR. For example, assuming that shiftAmount is 4, the
-     * following is true about AND and OR masks:
-     *
-     * <p>{@code expr & 0x0[anything] == 0x0[something] ;}
-     *
-     * <p>{@code expr | 0xF[anything] == 0xF[something] ;}
-     *
-     * @param maskKind the kind of mask (AND or OR)
-     * @param shiftAmountLit the LiteralTree whose value is shiftAmount
-     * @param maskLit the LiteralTree whose value is mask
-     * @param shiftedTypeKind the type of shift operation; int or long
-     * @return true iff the shiftAmount most significant bits of mask are 0 for AND, and 1 for OR
-     */
-    private boolean maskIgnoresMSB(
-            Kind maskKind,
-            LiteralTree shiftAmountLit,
-            LiteralTree maskLit,
-            TypeKind shiftedTypeKind) {
-        long shiftAmount = getLong(shiftAmountLit.getValue());
-
-        // Shift of zero is a nop
-        if (shiftAmount == 0) {
-            return true;
-        }
-
-        long mask = getLong(maskLit.getValue());
-        // Shift the shiftAmount most significant bits to become the shiftAmount least significant
-        // bits, zeroing out the rest.
-        if (shiftedTypeKind == TypeKind.INT) {
-            mask <<= 32;
-        }
-        mask >>>= (64 - shiftAmount);
-
-        if (maskKind == Kind.AND) {
-            // Check that the shiftAmount most significant bits of the mask were 0.
-            return mask == 0;
-        } else if (maskKind == Kind.OR) {
-            // Check that the shiftAmount most significant bits of the mask were 1.
-            return mask == (1 << shiftAmount) - 1;
-        } else {
-            throw new BugInCF("Invalid Masking Operation");
-        }
-    }
-
-    /**
-     * Given a casted right shift of the form {@code (type) (baseExpr >> shiftAmount)} or {@code
-     * (type) (baseExpr >>> shiftAmount)}, return true iff the expression's value is the same
-     * regardless of the type of right shift (signed or unsigned). This is true if the cast ignores
-     * the shiftAmount most significant bits of the shift result -- that is, if the cast ignores all
-     * the new bits that the right shift introduced on the left.
-     *
-     * <p>For example, the function returns true for
-     *
-     * <pre>{@code (short) (myInt >> 16)}</pre>
-     *
-     * and for
-     *
-     * <pre>{@code (short) (myInt >>> 16)}</pre>
-     *
-     * because these two expressions are guaranteed to have the same result.
-     *
-     * @param shiftTypeKind the kind of the type of the shift literal (BYTE, CHAR, SHORT, INT, or
-     *     LONG)
-     * @param castTypeKind the kind of the cast target type (BYTE, CHAR, SHORT, INT, or LONG)
-     * @param shiftAmountLit the LiteralTree whose value is shiftAmount
-     * @return true iff introduced bits are discarded
-     */
-    private boolean castIgnoresMSB(
-            TypeKind shiftTypeKind, TypeKind castTypeKind, LiteralTree shiftAmountLit) {
-        // Determine number of bits in the shift type, note shifts upcast to int.
-        // Also determine the shift amount as it is dependent on the shift type.
-        long shiftBits;
-        long shiftAmount;
-        switch (shiftTypeKind) {
-            case INT:
-                shiftBits = 32;
-                // When the LHS of the shift is an int, the 5 lower order bits of the RHS are used.
-                shiftAmount = 0x1F & getLong(shiftAmountLit.getValue());
-                break;
-            case LONG:
-                shiftBits = 64;
-                // When the LHS of the shift is a long, the 6 lower order bits of the RHS are used.
-                shiftAmount = 0x3F & getLong(shiftAmountLit.getValue());
-                break;
-            default:
-                throw new BugInCF("Invalid shift type");
-        }
-
-        // Determine number of bits in the cast type
-        long castBits;
-        switch (castTypeKind) {
-            case BYTE:
-                castBits = 8;
-                break;
-            case CHAR:
-                castBits = 8;
-                break;
-            case SHORT:
-                castBits = 16;
-                break;
-            case INT:
-                castBits = 32;
-                break;
-            case LONG:
-                castBits = 64;
-                break;
-            default:
-                throw new BugInCF("Invalid cast target");
-        }
-
-        long bitsDiscarded = shiftBits - castBits;
-
-        return shiftAmount <= bitsDiscarded || shiftAmount == 0;
-    }
-
-    /**
-     * Determines if a right shift operation, {@code >>} or {@code >>>}, is masked with a masking
-     * operation of the form {@code shiftExpr & maskLit} or {@code shiftExpr | maskLit} such that
-     * the mask renders the shift signedness ({@code >>} vs {@code >>>}) irrelevent by destroying
-     * the bits duplicated into the shift result. For example, the following pairs of right shifts
-     * on {@code byte b} both produce the same results under any input, because of their masks:
-     *
-     * <p>{@code (b >> 4) & 0x0F == (b >>> 4) & 0x0F;}
-     *
-     * <p>{@code (b >> 4) | 0xF0 == (b >>> 4) | 0xF0;}
-     *
-     * @param shiftExpr a right shift expression: {@code expr1 >> expr2} or {@code expr1 >>> expr2}
-     * @return true iff the right shift is masked such that a signed or unsigned right shift has the
-     *     same effect
-     */
-    private boolean isMaskedShiftEitherSignedness(BinaryTree shiftExpr) {
-        Pair<Tree, Tree> enclosingPair = TreeUtils.enclosingNonParen(visitorState.getPath());
-        // enclosing immediately contains shiftExpr or a parenthesized version of shiftExpr
-        Tree enclosing = enclosingPair.first;
-        // enclosingChild is a child of enclosing:  shiftExpr or a parenthesized version of it.
-        @SuppressWarnings("interning:assignment.type.incompatible") // comparing AST nodes
-        @InternedDistinct Tree enclosingChild = enclosingPair.second;
-
-        if (!isMask(enclosing)) {
-            return false;
-        }
-
-        BinaryTree maskExpr = (BinaryTree) enclosing;
-        ExpressionTree shiftAmountExpr = shiftExpr.getRightOperand();
-
-        // Determine which child of maskExpr leads to shiftExpr. The other one is the mask.
-        ExpressionTree mask =
-                maskExpr.getRightOperand() == enclosingChild
-                        ? maskExpr.getLeftOperand()
-                        : maskExpr.getRightOperand();
-
-        // Strip away the parentheses from the mask if any exist
-        mask = TreeUtils.withoutParens(mask);
-
-        if (!isLiteral(shiftAmountExpr) || !isLiteral(mask)) {
-            return false;
-        }
-
-        LiteralTree shiftLit = (LiteralTree) shiftAmountExpr;
-        LiteralTree maskLit = (LiteralTree) mask;
-
-        return maskIgnoresMSB(
-                maskExpr.getKind(), shiftLit, maskLit, TreeUtils.typeOf(shiftExpr).getKind());
-    }
-
-    /**
-     * Determines if a right shift operation, {@code >>} or {@code >>>}, is type casted such that
-     * the cast renders the shift signedness ({@code >>} vs {@code >>>}) irrelevent by discarding
-     * the bits duplicated into the shift result. For example, the following pair of right shifts on
-     * {@code short s} both produce the same results under any input, because of type casting:
-     *
-     * <p>{@code (byte)(s >> 8) == (byte)(b >>> 8);}
-     *
-     * @param shiftExpr a right shift expression: {@code expr1 >> expr2} or {@code expr1 >>> expr2}
-     * @return true iff the right shift is type casted such that a signed or unsigned right shift
-     *     has the same effect
-     */
-    private boolean isCastedShiftEitherSignedness(BinaryTree shiftExpr) {
-        // enclosing immediately contains shiftExpr or a parenthesized version of shiftExpr
-        Tree enclosing = TreeUtils.enclosingNonParen(visitorState.getPath()).first;
-
-        PrimitiveTypeTree castPrimitiveType = primitiveTypeCast(enclosing);
-        if (castPrimitiveType == null) {
-            return false;
-        }
-        TypeKind castTypeKind = castPrimitiveType.getPrimitiveTypeKind();
-
-        // Determine the type of the shift result
-        TypeKind shiftTypeKind =
-                atypeFactory.getAnnotatedType(shiftExpr).getUnderlyingType().getKind();
-
-        // Determine shift literal
-        ExpressionTree shiftAmountExpr = shiftExpr.getRightOperand();
-        if (!isLiteral(shiftAmountExpr)) {
-            return false;
-        }
-        LiteralTree shiftLit = (LiteralTree) shiftAmountExpr;
-
-        return castIgnoresMSB(shiftTypeKind, castTypeKind, shiftLit);
     }
 
     /**
@@ -348,16 +86,16 @@ public class SignednessVisitor extends BaseTypeVisitor<SignednessAnnotatedTypeFa
 
             case RIGHT_SHIFT:
                 if (hasUnsignedAnnotation(leftOpType)
-                        && !isMaskedShiftEitherSignedness(node)
-                        && !isCastedShiftEitherSignedness(node)) {
+                        && !atypeFactory.isMaskedShiftEitherSignedness(node, getCurrentPath())
+                        && !atypeFactory.isCastedShiftEitherSignedness(node, getCurrentPath())) {
                     checker.reportError(leftOp, "shift.signed", kind, leftOpType, rightOpType);
                 }
                 break;
 
             case UNSIGNED_RIGHT_SHIFT:
                 if (hasSignedAnnotation(leftOpType)
-                        && !isMaskedShiftEitherSignedness(node)
-                        && !isCastedShiftEitherSignedness(node)) {
+                        && !atypeFactory.isMaskedShiftEitherSignedness(node, getCurrentPath())
+                        && !atypeFactory.isCastedShiftEitherSignedness(node, getCurrentPath())) {
                     checker.reportError(leftOp, "shift.unsigned", kind, leftOpType, rightOpType);
                 }
                 break;
